# backend-lecture-code/day17-microservices/microservices/gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Day 17 Store — Microservices (via the API Gateway)
  version: 1.0.0
  description: |
    The same store as the monolith, split into eight services behind one gateway. The public
    paths are identical to the monolith's; only the port changed (4100 → 8080).

    **Log in first:** call `POST /users/login` with `{"email": "alice@shop.com"}`, copy the
    `token`, then click **Authorize** and paste it. The gateway checks the token once and passes
    the caller's id to the service in an `x-user-id` header. Services never see the token.

    Any routed request can also fail at the gateway itself: **502** when the service is down,
    **504** when it takes longer than 10 seconds.

    The **internal** section at the bottom lists service-to-service endpoints. The gateway has no
    route for them; they're documented with each service's own address for reference.

    Seed data to know: the 85" TV (`SKU-2001`, $5,999) is over the $5,000 limit, so paying for
    it is always declined. The Coffee Mug (`SKU-3001`) has only 1 in stock.
servers:
  - url: http://localhost:8080
    description: API gateway
...
paths:
...
  /users:
    post:
      tags: [users]
      summary: Sign up
      security: [] # public, no token needed
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NewUser" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /users/login:
    post:
      tags: [users]
      summary: Log in (demo — email only, no password)
      security: [] # public, no token needed
      description: The users service signs the JWT; the gateway verifies it on later requests.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: A JWT, valid for one hour
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /users/me:
    get:
      tags: [users]
      summary: The logged-in user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /orders:
    post:
      tags: [orders]
      summary: Check out my cart
      description: |
        The orders service calls five other services, in order:

        1. **cart** and **users** (in parallel): the items and the customer's email
        2. **catalog**: current prices, copied into the order as a snapshot
        3. **inventory**: reserve stock (committed on inventory's side immediately)
        4. saves the order in its own database as `pending_payment`
        5. **payments**: charge the card. If this fails or times out (3s), orders **releases the
           reservation itself** and marks the order `payment_failed`
        6. **cart**: empty it (best effort)
        7. **notifications**: fire-and-forget. If it's down, the order still succeeds.
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402":
          description: Card declined (any total over $5,000). The stock was released and the order is saved as `payment_failed`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamError" }
        "409":
          description: Not enough stock
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamError" }
        "502": { $ref: "#/components/responses/ServiceDown" }
        "503":
          description: A service that checkout depends on is unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamError" }
        "504":
          description: A service that checkout depends on timed out (e.g. payments after 3s)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamError" }
    get:
      tags: [orders]
      summary: My orders
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /payments/{id}:
    get:
      tags: [payments]
      summary: One of my payments
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /notifications:
    get:
      tags: [notifications]
      summary: Emails sent to me
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Email" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /notifications/debug/crash:
    post:
      tags: [demo]
      summary: Crash the notifications service (and only it)
      security: [] # public, no token needed
      description: |
        The same bug as the monolith's. It kills the notifications process only: `GET /health`
        shows just that one service down, and checkout still returns 201.
      responses:
        "200":
          description: The crash is scheduled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /payments/debug/slow:
    post:
      tags: [demo]
      summary: Make every card charge take this long
      security: [] # public, no token needed
      description: |
        Set `ms` above 3000 and checkout fails with **504** (orders stops waiting after 3
        seconds) and releases the stock. Set it back to `0` afterwards.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ms: { type: integer, example: 5000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  slowMs: { type: integer, example: 5000 }
        "502": { $ref: "#/components/responses/ServiceDown" }

  # ---- Internal: service to service, not routed by the gateway ----
  /internal/users/{id}:
    get:
      tags: [internal]
      summary: Look up any user by id (called by orders)
      security: [] # public, no token needed
      servers: [{ url: http://localhost:4101, description: users service }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /internal/payments/charges:
    post:
      tags: [internal]
      summary: Charge a card (called by orders)
      security: [] # public, no token needed
      servers: [{ url: http://localhost:4106, description: payments service }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, userId, amount]
              properties:
                orderId: { type: integer, example: 1 }
                userId: { type: integer, example: 1 }
                amount: { type: number, example: 194.48 }
      responses:
        "201":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "402":
          description: Declined (over $5,000). The declined payment is still recorded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  payment: { $ref: "#/components/schemas/Payment" }
  /internal/notifications/emails:
    post:
      tags: [internal]
      summary: Send an email (called by orders, fire-and-forget)
      security: [] # public, no token needed
      servers: [{ url: http://localhost:4107, description: notifications service }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, to, subject]
              properties:
                userId: { type: integer, example: 1 }
                to: { type: string, example: alice@shop.com }
                subject: { type: string, example: "Order #1 confirmed" }
                body: { type: string, example: "Total $194.48" }
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Email" }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, example: 1 }
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, or expired token (checked at the gateway), or unknown email on login
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Already exists, or not enough stock
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServiceDown:
      description: The service behind this route is unreachable (from the gateway itself)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UpstreamError" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: "unknown sku NOPE" }
    UpstreamError:
      type: object
      properties:
        error: { type: string, example: "payments timed out after 3000ms" }
        service: { type: string, example: payments }
    Message:
      type: object
      properties:
        message: { type: string }
...
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        email: { type: string, example: alice@shop.com }
        name: { type: string, example: Alice }
    NewUser:
      type: object
      required: [email, name]
      properties:
        email: { type: string, example: dave@shop.com }
        name: { type: string, example: Dave }
    LoginRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, example: alice@shop.com }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
...
    Order:
      type: object
      description: Includes a snapshot of the customer and items as they were at checkout.
      properties:
        id: { type: integer, example: 1 }
        userId: { type: integer, example: 1 }
        status:
          type: string
          enum: [pending_payment, placed, payment_failed]
          example: placed
        total: { type: number, example: 194.48 }
        customerEmail: { type: string, example: alice@shop.com }
        items:
          type: array
          items:
            type: object
            properties:
              sku: { type: string, example: SKU-1001 }
              name: { type: string, example: Running Shoes }
              quantity: { type: integer, example: 2 }
              unitPrice: { type: number, example: 89.99 }
        reservationId: { type: integer, example: 1 }
        paymentId: { type: integer, nullable: true, example: 1 }
        createdAt: { type: string, format: date-time }
    Payment:
      type: object
      properties:
        id: { type: integer, example: 1 }
        orderId: { type: integer, example: 1 }
        userId: { type: integer, example: 1 }
        amount: { type: number, example: 194.48 }
        status: { type: string, enum: [approved, declined], example: approved }
    Email:
      type: object
      properties:
        id: { type: integer, example: 1 }
        userId: { type: integer, example: 1 }
        to: { type: string, example: alice@shop.com }
        subject: { type: string, example: "Order #1 confirmed" }
        body: { type: string, example: "Total $194.48" }
        sentAt: { type: string, format: date-time }
...